Support EasyBuild 4.4.x, Python3 only, LMOD 8.5.x

+++ run_depoly.sh
@@ -1,7 +1,9 @@
 #!/bin/bash
 
+[[ -z "${TOOLCHAIN}" ]] && { echo 'TOOLCHAIN not set' ; exit; }
+[[ -z "${EB_VER}" ]] && { echo 'EB_VER not set' ; exit; }
 
 docker run --rm -ti --name ${TOOLCHAIN#"foss-"} \
- --security-opt="seccomp:unconfined" \
+ --security-opt="seccomp=unconfined" \
  -v /app:/app \
  fredhutch/ls2:eb-${EB_VER}-${TOOLCHAIN} bash
